chore(onboarding): create secure onboarding client
