+++ sysdig/internal/client/v2/cloud_account.go
@@ -11,8 +11,6 @@ const (
 	cloudAccountsWithExternalIDPath = "%s/api/cloud/v2/accounts?includeExternalID=true&upsert=true"
 	cloudAccountPath                = "%s/api/cloud/v2/accounts/%s"
 	cloudAccountWithExternalIDPath  = "%s/api/cloud/v2/accounts/%s?includeExternalID=true"
-	onboardingTrustedIdentityPath   = "%s/api/secure/onboarding/v2/trustedIdentity?provider=%s"
-	onboardingTenantExternaIDPath   = "%s/api/secure/onboarding/v2/externalID"
 	providersPath                   = "%v/api/v2/providers"
 )
 
@@ -22,8 +20,6 @@ type CloudAccountSecureInterface interface {
 	GetCloudAccountSecure(ctx context.Context, accountID string) (*CloudAccountSecure, error)
 	DeleteCloudAccountSecure(ctx context.Context, accountID string) error
 	UpdateCloudAccountSecure(ctx context.Context, accountID string, cloudAccount *CloudAccountSecure) (*CloudAccountSecure, error)
-	GetTrustedCloudIdentitySecure(ctx context.Context, provider string) (string, error)
-	GetTenantExternalIDSecure(ctx context.Context) (string, error)
 }
 
 type CloudAccountMonitorInterface interface {
@@ -99,34 +95,6 @@ func (client *Client) UpdateCloudAccountSecure(ctx context.Context, accountID st
 	}
 
 	return Unmarshal[*CloudAccountSecure](response.Body)
-}
-
-func (client *Client) GetTrustedCloudIdentitySecure(ctx context.Context, provider string) (string, error) {
-	response, err := client.requester.Request(ctx, http.MethodGet, fmt.Sprintf(onboardingTrustedIdentityPath, client.config.url, provider), nil)
-	if err != nil {
-		return "", err
-	}
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusOK {
-		return "", client.ErrorFromResponse(response)
-	}
-
-	return Unmarshal[string](response.Body)
-}
-
-func (client *Client) GetTenantExternalIDSecure(ctx context.Context) (string, error) {
-	response, err := client.requester.Request(ctx, http.MethodGet, fmt.Sprintf(onboardingTenantExternaIDPath, client.config.url), nil)
-	if err != nil {
-		return "", err
-	}
-	defer response.Body.Close()
-
-	if response.StatusCode != http.StatusOK {
-		return "", client.ErrorFromResponse(response)
-	}
-
-	return Unmarshal[string](response.Body)
 }
 
 func (client *Client) cloudAccountsURL(includeExternalID bool) string {
